Renderable/DisplayerImpl: Fix bitmap fuzzy issue

+++ src/Displayer.cpp
@@ -41,7 +41,7 @@ namespace WTFDanmaku {
         pImpl->Resize(width, height);
     }
 
-    ComPtr<ID2D1Bitmap1> Displayer::CreateBitmap(float width, float height) {
+    ComPtr<ID2D1Bitmap1> Displayer::CreateBitmap(uint32_t width, uint32_t height) {
         return pImpl->CreateBitmap(width, height);
     }
